add model

+++ domin/util/type.go
@@ -3,7 +3,7 @@ package util
 import "github.com/dgrijalva/jwt-go"
 
 type tokenClaims struct {
-	UserID    uint64 `json:"UserID"`
+	UserID    int32  `json:"UserID"`
 	FirstName string `json:"FirstName"`
 	LastName  string `json:"LastName"`
 	Email     string `json:"Email"`
